+++ src/components/HadithWidget.test.js
@@ -2,6 +2,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { MemoryRouter } from "react-router";
 import { HadithWidget } from "./HadithWidget";
+import { I18nextProvider } from "react-i18next";
+import { i18n } from "../i18n-test";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");
@@ -13,7 +15,9 @@ it("renders without crashing", () => {
 
   ReactDOM.render(
     <MemoryRouter>
-      <HadithWidget hadith={hadith} />
+      <I18nextProvider i18n={i18n}>
+        <HadithWidget hadith={hadith} />
+      </I18nextProvider>
     </MemoryRouter>,
     div
   );
