--- src/containers/HadithsPage/test.js
@@ -1,9 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { HadithsPage } from "./";
-
-it("renders without crashing", () => {
-  const div = document.createElement("div");
-  ReactDOM.render(<HadithsPage />, div);
-  ReactDOM.unmountComponentAtNode(div);
-});
