+++ src/index.js
@@ -5,47 +5,14 @@ import "./index.css";
 import "typeface-roboto";
 import * as serviceWorker from "./serviceWorker";
 import { Layout } from "./containers/Layout/";
-import i18n from "i18next";
-import Backend from "i18next-xhr-backend";
-import LanguageDetector from "i18next-browser-languagedetector";
-import { reactI18nextModule } from "react-i18next";
-
-function formatNumber(number, language) {
-  return number.toLocaleString(language);
-}
-
-i18n
-  // detect user language
-  // learn more: https://github.com/i18next/i18next-browser-languageDetector
-  .use(LanguageDetector)
-  // load translation using xhr -> see /public/locales
-  // learn more: https://github.com/i18next/i18next-xhr-backend
-  .use(Backend)
-  // pass the i18n instance to react-i18next.
-  .use(reactI18nextModule)
-  // init i18next
-  // for all options read: https://www.i18next.com/overview/configuration-options
-  .init({
-    fallbackLng: "en",
-    debug: true,
-    keySeparator: ".", // we do not use keys in form messages.welcome
-    interpolation: {
-      format: function(value, format, language) {
-        if (format === "number") {
-          return formatNumber(value, language);
-        }
-        return value;
-      },
-      escapeValue: false // not needed for react as it escapes by default
-    },
-    react: {
-      wait: true
-    }
-  });
+import { I18nextProvider } from "react-i18next";
+import { i18n } from "./i18n";
 
 ReactDOM.render(
   <BrowserRouter>
-    <Layout />
+    <I18nextProvider i18n={i18n}>
+      <Layout />
+    </I18nextProvider>
   </BrowserRouter>,
   document.getElementById("root")
 );
